% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGeneInfoFromGtfToResTab.R
\name{addGeneInfoFromGtfToResTab}
\alias{addGeneInfoFromGtfToResTab}
\title{Add gene information from GTF to given data frame}
\usage{
addGeneInfoFromGtfToResTab(res, gtf)
}
\arguments{
\item{res}{a data frame}

\item{gtf}{a GRanges object}
}
\value{
a data frame
}
\description{
Add gene information from GTF to given data frame
}
