% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAndCreateCorrectDesignString.R
\name{checkAndCreateCorrectDesignString}
\alias{checkAndCreateCorrectDesignString}
\title{Tests given design is correct}
\usage{
checkAndCreateCorrectDesignString(s_sheet, design)
}
\arguments{
\item{s_sheet}{a data frame}

\item{design}{a string}
}
\value{
a string
}
\description{
Tests given design is correct
}
