% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsDensityPlot.R
\name{countsDensityPlot}
\alias{countsDensityPlot}
\title{count density plot}
\usage{
countsDensityPlot(s_sheet, countsData)
}
\arguments{
\item{s_sheet}{a data frame; sample metadata sheet}

\item{countsData}{a matrix; raw counts matrix}
}
\value{
An object created by \code{ggplot}
}
\description{
count density plot
}
