% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get3dPCAplot.R
\name{get3dPCAplot}
\alias{get3dPCAplot}
\title{Gives 3d PCA plot}
\usage{
get3dPCAplot(countsDat, s_sheet, pcaColFactor = "SampleGroup")
}
\arguments{
\item{countsDat}{a counts matrix}

\item{s_sheet}{a data frame; sample sheet used for plot aesthetics}

\item{pcaColFactor}{a character vector with one value; column name in a sample sheet for PCA color aesthetic}
}
\value{
a plot object created by plot_ly
}
\description{
Gives 3d PCA plot
}
