% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneCountsPlot.R
\name{getGeneCountsPlot}
\alias{getGeneCountsPlot}
\title{Gives read counts scater plot for topN genes}
\usage{
getGeneCountsPlot(
  dds,
  factorName,
  numerator,
  denominator,
  gtf,
  pValCutoff = 0.05,
  genesToShow = NULL,
  topN = 6
)
}
\arguments{
\item{dds}{dds DESeqDataSet object; dds after applying \code{estimateSizeFactors}, \code{estimateDispersions} and \code{nbinomWaldTest}}

\item{factorName}{s string}

\item{numerator}{a string; numerator in a contrast}

\item{denominator}{a string; denominator (reference) in a contrast}

\item{gtf}{a GRanges object; typical output from \code{import.gff}}

\item{pValCutoff}{numeric value; padj cut off value to classify as significant}

\item{genesToShow}{a string; list genes to show}

\item{topN}{integer; number of top gene names to show on plot; default top 6}
}
\value{
An object created by \code{ggplot}
}
\description{
Gives read counts scater plot for topN genes
}
