% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneDetctionPlot.R
\name{getGeneDetctionPlot}
\alias{getGeneDetctionPlot}
\title{Gives gene detection plot}
\usage{
getGeneDetctionPlot(countsData, s_sheet, thresholds = seq(0, 50, by = 5))
}
\arguments{
\item{countsData}{a matrix; raw counts matrix}

\item{s_sheet}{a data frame; sample metadata sheet}

\item{thresholds}{a vector of counts thresholds}
}
\value{
An object created by \code{ggplot}
}
\description{
Gives gene detection plot
}
