% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHeatmap.R
\name{getHeatmap}
\alias{getHeatmap}
\title{Gives heat-map}
\usage{
getHeatmap(dds, topN = 500, annoGroup = "SampleGroup")
}
\arguments{
\item{dds}{DESeqDataSet object}

\item{topN}{integer; top N highly variable genes to use in heat-map}

\item{annoGroup}{a character vector; one more meta data column names for annotating heat-map. Default is SampleGroup}
}
\description{
Gives heat-map
}
