% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKaryogramPlot.R
\name{getKaryogramPlot}
\alias{getKaryogramPlot}
\title{Gives Karyogram Plot}
\usage{
getKaryogramPlot(
  dds,
  factorName,
  numerator,
  denominator,
  gtf,
  pValCutoff = 0.05,
  genome = NULL
)
}
\arguments{
\item{dds}{dds DESeqDataSet object; dds after applying \code{estimateSizeFactors}, \code{estimateDispersions} and \code{nbinomWaldTest}}

\item{factorName}{s string}

\item{numerator}{a string; numerator in a contrast}

\item{denominator}{a string; denominator (reference) in a contrast}

\item{gtf}{a GRanges object; typical output from \code{import.gff}}

\item{pValCutoff}{a numeric value}

\item{genome}{a string; species name like 'Homo_sapiens'}
}
\value{
An object created by \code{ggplot}
}
\description{
Gives Karyogram Plot
}
