% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLibraryComplexityPlot.R
\name{getLibraryComplexityPlot}
\alias{getLibraryComplexityPlot}
\title{Gives library complexity plot}
\usage{
getLibraryComplexityPlot(countsData, s_sheet)
}
\arguments{
\item{countsData}{a matrix; raw counts matrix}

\item{s_sheet}{a data frame; sample metadata sheet}
}
\value{
An object created by \code{ggplot}
}
\description{
Gives library complexity plot
}
