% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMAplot.R
\name{getMAplot}
\alias{getMAplot}
\title{returns MA plot}
\usage{
getMAplot(
  res,
  numerator,
  denominator,
  genesToShow = NULL,
  topN = 50,
  pValCutoff = 0.05
)
}
\arguments{
\item{res}{a data frame; DESeq2 results converted to data frame}

\item{numerator}{a string; numerator in a contrast}

\item{denominator}{a string; denominator in a contrast}

\item{genesToShow}{a string; list genes to show}

\item{topN}{integer; number of top gene names to show on plot; default top 50}

\item{pValCutoff}{numeric value; padj cut off value to classify as significant}
}
\value{
An object created by \code{ggplot}
}
\description{
returns MA plot
}
