% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVolcanoPlot.R
\name{getVolcanoPlot}
\alias{getVolcanoPlot}
\title{gives volcano plot}
\usage{
getVolcanoPlot(
  res,
  numerator,
  denominator,
  topN = 50,
  genesToShow = NULL,
  pValCutoff = 0.05
)
}
\arguments{
\item{res}{a data frame; DESeq2 results converted to data frame}

\item{numerator}{numerator  a string; numerator in a contrast}

\item{denominator}{denominator a string; denominator in a contrast}

\item{topN}{integer; number of top gene names to show on plot; default top 50}

\item{genesToShow}{a vector; list genes to show}

\item{pValCutoff}{numeric value; padj cut off value to classify as significant}
}
\value{
An object created by \code{ggplot}
}
\description{
gives volcano plot
}
