% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalClustPlot.R
\name{hierarchicalClustPlot}
\alias{hierarchicalClustPlot}
\title{Hierarchical clustering plot
Computes and displays hierarchical clustering plot for samples in a given counts data}
\usage{
hierarchicalClustPlot(
  countsDat,
  s_sheet,
  colorByCol = "SampleGroup",
  horizontal = TRUE,
  sampleColors = NULL,
  title = "",
  topN = 1000
)
}
\arguments{
\item{countsDat}{a matrix;  vst or rlog transformed counts matrix}

\item{s_sheet}{a data frame; sample metadata}

\item{colorByCol}{a character vector; name of sample metadata column to color the samples}

\item{horizontal}{a boolean value; orientation of the dendrogram. Default is TRUE.}

\item{sampleColors}{a vector of colors; if NULL, colors are automatically assigned.}

\item{title}{a character vector; the main title for the dendrogram.}

\item{topN}{an integer; top highly variable genes}
}
\value{
An object created by \code{ggplot}
}
\description{
Hierarchical clustering plot
Computes and displays hierarchical clustering plot for samples in a given counts data
}
