% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSampleTxiAndSaveRds.R
\name{importSampleTxiAndSaveRds}
\alias{importSampleTxiAndSaveRds}
\title{Save combined txi object after reading sample level txi data}
\usage{
importSampleTxiAndSaveRds(
  s_sheet,
  quantOut,
  tx2gene,
  quantTool = "salmon",
  save = TRUE
)
}
\arguments{
\item{s_sheet}{A data frame with at least one column named SampleName.}

\item{quantOut}{A character vector, salmon output path.}

\item{tx2gene}{A data frame with TxID (transcript ids) and GeneID columns.}

\item{quantTool}{A character vector, quantification tool, like salmon}

\item{save}{Boolean value, do you want to save combined txi object, default is true, saves 'txi.rds' in quantOut}
}
\value{
a txi object
}
\description{
Save combined txi object after reading sample level txi data
}
