% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_fileExists.R
\name{is_fileExists}
\alias{is_fileExists}
\title{tests if file exists}
\usage{
is_fileExists(filesToCheck)
}
\arguments{
\item{filesToCheck}{a vector of files to test}
}
\value{
Boolean value
}
\description{
tests if file exists
}
