% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_validColor.R
\name{is_validColor}
\alias{is_validColor}
\title{tests if color string is valid}
\usage{
is_validColor(color)
}
\arguments{
\item{color}{a string}
}
\value{
Boolean value
}
\description{
tests if color string is valid
}
