% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_validMetaData.R
\name{is_validMetaData}
\alias{is_validMetaData}
\title{check if the metadata is valid}
\usage{
is_validMetaData(s_sheet, columnsToCheck)
}
\arguments{
\item{s_sheet}{a data frame}

\item{columnsToCheck}{string; column name}
}
\value{
Boolean value
}
\description{
check if the metadata is valid
}
