% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_validSampleColors.R
\name{is_validSampleColors}
\alias{is_validSampleColors}
\title{Tests that given sample colors valid}
\usage{
is_validSampleColors(s_sheet, sampleColors, colorByCol)
}
\arguments{
\item{s_sheet}{data frame; sample metadata sheet}

\item{sampleColors}{colors vector}

\item{colorByCol}{column name}
}
\value{
Boolean value
}
\description{
Tests that given sample colors valid
}
