% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_validTx2gene.R
\name{is_validTx2gene}
\alias{is_validTx2gene}
\title{tests if tx2gene format is correct}
\usage{
is_validTx2gene(tx2gene)
}
\arguments{
\item{tx2gene}{a data frame}
}
\value{
Boolean value
}
\description{
tests if tx2gene format is correct
}
