% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneBiotypes.R
\name{plotGeneBiotypes}
\alias{plotGeneBiotypes}
\title{gives Gene biotypes read share}
\usage{
plotGeneBiotypes(countsData, gtf, s_sheet)
}
\arguments{
\item{countsData}{counts matrix}

\item{gtf}{GRanges object}

\item{s_sheet}{a data frame; sample metadata}
}
\value{
An object created by \code{ggplot}
}
\description{
gives Gene biotypes read share
}
