% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readsPerSampleBarPlot.R
\name{readsPerSampleBarPlot}
\alias{readsPerSampleBarPlot}
\title{barplot of read counts per sample.}
\usage{
readsPerSampleBarPlot(readCounts, s_sheet)
}
\arguments{
\item{readCounts}{a data frame with at least two columns SampleName and fragments}

\item{s_sheet}{a data frame; sample metadata with at lest two columns SampleName and SampleGroup}
}
\value{
a ggplot object
}
\description{
barplot of read counts per sample.
}
