% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveCounts.R
\name{saveCounts}
\alias{saveCounts}
\title{write counts data}
\usage{
saveCounts(dds, txi, countsDir, gtf)
}
\arguments{
\item{dds}{DESeqDataSet object; dds after applying estimateSizeFactors}

\item{txi}{a list; typical output of \code{tximport}}

\item{countsDir}{a string; directory to write counts files}

\item{gtf}{a GRanges object; typical output from \code{import.gff}}
}
\description{
write counts data
}
