% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformCounts.R
\name{transformCounts}
\alias{transformCounts}
\title{Filter and transform raw counts}
\usage{
transformCounts(rawCounts, countsCutOff = 10, FUN = vst)
}
\arguments{
\item{rawCounts}{a counts matrix}

\item{countsCutOff}{an integer, keep only rows with a total row count of at least countsCutOff}

\item{FUN}{a function name for transformation. vst is the default transformation function, but rlog, log2, etc., can also be selected.}
}
\value{
a transformed counts matrix
}
\description{
Filter and transform raw counts
}
