% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignColors.R
\name{assignColors}
\alias{assignColors}
\title{Assign colors to samples / sample groups}
\usage{
assignColors(s_sheet, colorByCol = "SampleGroup")
}
\arguments{
\item{s_sheet, }{a data frame; metadata sheet}

\item{colorByCol}{character vector of length 1; column name of the metadata sheet for color assignment.}
}
\value{
a named vector of colors
}
\description{
Assign colors to samples / sample groups
}
\examples{

}
