% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationPlot.R
\name{correlationPlot}
\alias{correlationPlot}
\title{Correlation plot from counts matrix}
\usage{
correlationPlot(countsMat)
}
\arguments{
\item{countsMat}{a counts matrix}
}
\value{

}
\description{
Correlation plot from counts matrix
}
\examples{

}
