% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationPlot.R
\name{correlationPlot}
\alias{correlationPlot}
\title{Correlation plot from counts matrix}
\usage{
correlationPlot(countsMat, s_sheet, topN = 1000)
}
\arguments{
\item{countsMat}{a counts matrix}

\item{s_sheet}{a data frame; sample meta data sheet}
}
\value{

}
\description{
Correlation plot from counts matrix
}
\examples{

}
