% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDdsAndDESeq.R
\name{createDdsAndDESeq}
\alias{createDdsAndDESeq}
\title{Create DDS object and do basic DE}
\usage{
createDdsAndDESeq(txi, s_sheet, design, fitType = "parametric")
}
\arguments{
\item{txi}{a txi object}

\item{s_sheet}{a data frame; sample metadata sheet}

\item{design}{a string; design}

\item{fitType}{a string; fit Type for dispersion estimates}
}
\value{
a DESeqDataSet object
}
\description{
Create DDS object and do basic DE
}
\examples{

}
