% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPcaLoadingsPlot.R
\name{getPcaLoadingsPlot}
\alias{getPcaLoadingsPlot}
\title{Gives PCA loadings rank plot}
\usage{
getPcaLoadingsPlot(countsData, s_sheet, gtf, topN = 10, genesToShow = NULL)
}
\arguments{
\item{countsData}{a counts matrix}

\item{s_sheet}{a data frame; sample sheet used for plot aesthetics}

\item{gtf}{a GRanges object; typical output from \code{import.gff}}

\item{topN}{integer; number of top gene names to show on plot; default top 10}

\item{genesToShow}{a string; list genes to show}
}
\value{
An object created by \code{ggplot}
}
\description{
Gives PCA loadings rank plot
}
\examples{

}
