% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPcaPlot.R
\name{getPcaPlot}
\alias{getPcaPlot}
\title{get PCA plots given counts matrix and sample sheet}
\usage{
getPcaPlot(
  countsDat,
  s_sheet,
  pcaColFactor = "SampleGroup",
  pcaShapeFac = NULL,
  PCx = 1,
  PCy = 2
)
}
\arguments{
\item{countsDat}{a counts matrix}

\item{s_sheet}{a data frame; sample sheet used for plot aesthetics}

\item{pcaColFactor}{a character vector with one value; column name in a sample sheet for PCA color aesthetic}

\item{pcaShapeFac}{a character vector with one value; column name in a sample sheet for PCA shape aesthetic}

\item{PCx}{an integer, X-axis PC, default is 1, that is PC1}

\item{PCy}{an integer, Y-axis PC, default is 2, that is PC2}
}
\value{
a plot
}
\description{
get PCA plots given counts matrix and sample sheet
}
\examples{

}
