% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPcaPlot.R
\name{getPcaPlot}
\alias{getPcaPlot}
\title{get PCA plots given counts matrix and sample sheet}
\usage{
getPcaPlot(
  countsDat,
  s_sheet,
  pcaColFactor = "SampleGroup",
  pcaShapeFac = NULL
)
}
\arguments{
\item{countsDat}{a counts matrix}

\item{s_sheet}{a data frame; sample sheet used for plot aesthetics}

\item{pcaColFactor}{a character vector with one value; column name in a sample sheet for PCA color aesthetic}

\item{pcaShapeFac}{a character vector with one value; column name in a sample sheet for PCA shape aesthetic}
}
\description{
get PCA plots given counts matrix and sample sheet
}
\examples{

}
