% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPvalDistriPlot.R
\name{getPvalDistriPlot}
\alias{getPvalDistriPlot}
\title{Gives p value distribution plot}
\usage{
getPvalDistriPlot(res, numerator, denominator, pValCutoff)
}
\arguments{
\item{res}{a data frame; DESeq2 results converted to data frame}

\item{numerator}{a string; numerator in a contrast}

\item{denominator}{a string; denominator in a contrast}

\item{pValCutoff}{numeric value; padj cut off value}
}
\value{
An object created by \code{ggplot}
}
\description{
Gives p value distribution plot
}
\examples{

}
