% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopVarGenes.R
\name{getTopVarGenes}
\alias{getTopVarGenes}
\title{Gives top highly variable genes}
\usage{
getTopVarGenes(countsDat, topN = 1000)
}
\arguments{
\item{countsDat}{a matrix}

\item{topN}{an integer}
}
\value{
a matrix
}
\description{
Gives top highly variable genes
}
\examples{

}
