% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normFactorsBoxplot.R
\name{normFactorsBoxplot}
\alias{normFactorsBoxplot}
\title{gives normalization factors, raw and normalized counts distribution box plots}
\usage{
normFactorsBoxplot(dds)
}
\arguments{
\item{dds}{a DESeqDataSet object; dds after applying estimateSizeFactors}
}
\value{
An object created by \code{ggplot}
}
\description{
gives normalization factors, raw and normalized counts distribution box plots
}
\examples{

}
