% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneBiotypes.R
\name{plotGeneBiotypes}
\alias{plotGeneBiotypes}
\title{gives Gene biotypes read share}
\usage{
plotGeneBiotypes(dds, gtf)
}
\arguments{
\item{dds}{DESeqDataSet object}

\item{gtf}{GRanges object}
}
\value{
An object created by \code{ggplot}
}
\description{
gives Gene biotypes read share
}
\examples{

}
