% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readsPerGroupBoxplot.R
\name{readsPerGroupBoxplot}
\alias{readsPerGroupBoxplot}
\title{Sample group read counts distribution box plot}
\usage{
readsPerGroupBoxplot(s_sheet, readCounts)
}
\arguments{
\item{s_sheet}{a data frame; sample metadata with at lest two columns SampleName and SampleGroup}

\item{readCounts}{a data frame with at least two columns SampleName and fragments}
}
\value{
An object created by \code{ggplot}
}
\description{
Sample group read counts distribution box plot
}
\examples{

}
