% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboFractionPlot.R
\name{riboFractionPlot}
\alias{riboFractionPlot}
\title{Get Ribosomal and Nonribosomal counts barplot.}
\usage{
riboFractionPlot(countsData, gtf)
}
\arguments{
\item{countsData}{a matrix; counts matrix preferably raw counts matrix.}

\item{gtf}{a GRanges object; gft}
}
\value{
An object created by \code{ggplot}
}
\description{
Get Ribosomal and Nonribosomal counts barplot.
}
