% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeAndPlotDEResults.R
\name{writeAndPlotDEResults}
\alias{writeAndPlotDEResults}
\title{write DE results and create plots for each contrast}
\usage{
writeAndPlotDEResults(dds, factorName, numerator, denominator, DeOutDir, gtf)
}
\arguments{
\item{dds}{dds DESeqDataSet object; dds after applying \code{estimateSizeFactors}, \code{estimateDispersions} and \code{nbinomWaldTest}}

\item{factorName}{s string}

\item{numerator}{a string; numerator in a contrast}

\item{denominator}{a string; denominator (reference) in a contrast}

\item{DeOutDir}{a string; directory name}

\item{gtf}{a GRanges object; typical output from \code{import.gff}}
}
\value{

}
\description{
write DE results and create plots for each contrast
}
\examples{

}
